Adjust signature of generateRSAKey

+++ overlord/devicestate/export_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/snapcore/snapd/overlord/state"
 )
 
-func MockKeyLength(n uint64) (restore func()) {
+func MockKeyLength(n int) (restore func()) {
 	oldKeyLength := keyLength
 	if n < 1024 {
 		n = 1024
